+++ mini_socket/src/os_epoll.rs
@@ -32,7 +32,7 @@ impl OSEpoll {
     }
 
     #[inline]
-    pub fn ctl_add_fd(&self, id: u64, fd: RawFd, ev: i32) -> Result<(), String> {
+    pub fn ctl_add_fd(&self, id: u32, fd: RawFd, ev: i32) -> Result<(), String> {
         let mut event = libc::epoll_event {
             u64: (id as libc::c_ulonglong),
             events: (EPOLL_EVENTS | ev) as u32,
@@ -46,7 +46,7 @@ impl OSEpoll {
         }
     }
     #[inline]
-    pub fn ctl_mod_fd(&self, id: u64, fd: RawFd, ev: i32) -> Result<(), String> {
+    pub fn ctl_mod_fd(&self, id: u32, fd: RawFd, ev: i32) -> Result<(), String> {
         let mut event = libc::epoll_event {
             u64: (id as libc::c_ulonglong),
             events: (EPOLL_EVENTS | ev) as u32,
@@ -60,7 +60,7 @@ impl OSEpoll {
         }
     }
     #[inline]
-    pub fn ctl_del_fd(&self, id: u64, fd: RawFd) -> Result<(), String> {
+    pub fn ctl_del_fd(&self, id: u32, fd: RawFd) -> Result<(), String> {
         let mut event = libc::epoll_event {
             events: 0,
             u64: (id as libc::c_ulonglong),
